backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0
pdf2image>=1.16.0
pillow>=9.0.0
openpyxl>=3.0.0
python-multipart>=0.0.5